requirements.txt
streamlit
torch
numpy
transformers
langchain_groq
python-dotenv
llama_parse
langchain
llama_index
python-multipart

